+++ chunkflow/flow/plugin.py
@@ -1,11 +1,14 @@
 # -*- coding: utf-8 -*-
 import os.path as path
 
+import numpy as np
+
 from .base import OperatorBase
 
 from cloudvolume.lib import Bbox
 
 from chunkflow.lib import load_source
+from chunkflow.chunk import Chunk
 
 
 class Plugin(OperatorBase):
@@ -41,13 +44,36 @@ class Plugin(OperatorBase):
         program = load_source(plugin_file)
         
         # assuming this is a func / static functor for now, maybe make it a class?
-        self.exec = program.exec  
+        self.execute = program.execute
 
-    def __call__(self, inputs, bbox: Bbox = None, args: str = None):
-        if args is None and bbox is None:
-            return self.exec(*inputs)
-        if args is None and bbox is not None:
-            return self.exec(*inputs, bbox=bbox)
-        if args is not None and bbox is None:
-            return self.exec(*inputs, args=args) 
-        return self.exec(*inputs, bbox=bbox, args=args)
+    def __call__(self, inputs: list, args: str = None):
+        voxel_offset = None
+        voxel_size = None
+        shape = None
+        for inp in inputs:
+            if isinstance(inp, Chunk):
+                voxel_offset = inp.voxel_offset
+                voxel_size = inp.voxel_size
+                shape = inp.shape
+                break
+
+        if len(inputs) == 0 and args is None:
+            outputs = self.execute()
+        elif len(inputs) == 0 and args is not None:
+            outputs = self.execute(args=args)
+        elif len(inputs) > 0 and args is None:
+            outputs = self.execute(*inputs)
+        else:
+            outputs = self.execute(*inputs, args=args) 
+
+        # automatically convert the ndarrays to Chunks
+        if voxel_offset is not None:
+            assert voxel_size is not None
+            assert shape is not None
+            for idx, output in enumerate(outputs):
+                if isinstance(output, np.ndarray):
+                    # in case the plugin did some symmetric cropping
+                    offset = tuple(vo + (ins - outs)//2 for vo, ins, outs in zip(voxel_offset, shape[-3:], output.shape[-3:]) )
+                    outputs[idx] = Chunk(output, voxel_offset=offset, voxel_size=voxel_size)
+        
+        return outputs